Adds 'main' file
This commit adds a main file. This is where the script actually runs, pulling in all the functions previously written. It asks the user to confirm what they'd like to do and gives them an opportunity to back out. It also assumes that the username and password are stored as environment variables, rather than being written directly into the script. If they're not in the environment, the script will ask for them.